backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
Pillow==11.0.0
opencv-python==4.10.0.84
torch==2.5.1
torchvision==0.20.1
reportlab==4.2.5
